Fixed style

+++ test/regression/factory.cc
@@ -101,8 +101,8 @@ TEST_F(FactoryTest, BlackCamera)
   unsigned int width;
   unsigned int height;
   GetFrame("camera_sensor", &img, width, height);
-  ASSERT_EQ(width, 320);
-  ASSERT_EQ(height, 240);
+  ASSERT_EQ(width, static_cast<unsigned int>(320));
+  ASSERT_EQ(height, static_cast<unsigned int>(240));
 
   unsigned char *trueImage = new unsigned char[width * height * 3];
   memset(trueImage, 178, width*height*3);
@@ -111,8 +111,8 @@ TEST_F(FactoryTest, BlackCamera)
   unsigned int diffSum = 0;
   double diffAvg = 0;
   ImageCompare(&img, &trueImage, width, height, 3, diffMax, diffSum, diffAvg);
-  ASSERT_EQ(diffSum, 0);
-  ASSERT_EQ(diffMax, 0);
+  ASSERT_EQ(diffSum, static_cast<unsigned int>(0));
+  ASSERT_EQ(diffMax, static_cast<unsigned int>(0));
   ASSERT_EQ(diffAvg, 0.0);
 }
 
@@ -130,8 +130,8 @@ TEST_F(FactoryTest, Camera)
   unsigned int width;
   unsigned int height;
   GetFrame("camera_sensor2", &img, width, height);
-  ASSERT_EQ(width, 320);
-  ASSERT_EQ(height, 240);
+  ASSERT_EQ(width, static_cast<unsigned int>(320));
+  ASSERT_EQ(height, static_cast<unsigned int>(240));
 
   unsigned int diffMax = 0;
   unsigned int diffSum = 0;
@@ -139,8 +139,8 @@ TEST_F(FactoryTest, Camera)
   ImageCompare(&img, &empty_world_camera1,
       width, height, 3, diffMax, diffSum, diffAvg);
   // PrintImage("empty_world_camera1", &img, width, height, 3);
-  ASSERT_EQ(diffSum, 0);
-  ASSERT_EQ(diffMax, 0);
+  ASSERT_EQ(diffSum, static_cast<unsigned int>(0));
+  ASSERT_EQ(diffMax, static_cast<unsigned int>(0));
   ASSERT_EQ(diffAvg, 0.0);
 }
